Criado controller e service de Cidade. Criando o CRUD e implementando os tratamentos de erros. Implementando o método deletar no Restaurante Controller e fazendo os tratamentos de erros. Ajustando o método buscar no controller do Estado

+++ src/main/java/com/algaworks/algafood/infrastructure/repository/RestauranteRepopositoryImpl.java
@@ -5,6 +5,7 @@ import java.util.List;
 import javax.persistence.EntityManager;
 import javax.persistence.PersistenceContext;
 
+import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.stereotype.Component;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -35,8 +36,13 @@ public class RestauranteRepopositoryImpl implements RestauranteRepository {
 
 	@Override
 	@Transactional
-	public void remover(Restaurante restaurante) {
-		restaurante = buscar(restaurante.getId());
+	public void remover(Long id) {
+		Restaurante restaurante = buscar(id);
+		
+		if (restaurante == null) {
+			throw new EmptyResultDataAccessException(1);
+		}
+		
 		entityManager.remove(restaurante);
 	}
